[obey-campaign:8deed8b4] test(triage): keep refresh test files under the size standard
Splits the append and remote/carry cases out of refresh_test.go, which
crossed 500 lines as the review regression tests landed. No assertions
changed; largest test file is now 463.
